requirements.txt
kafka-python>=0.9.4
